pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "colorama>=0.4.6",
    "dnspython>=2.8.0",
    "python-whois>=0.9.6",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
]
